// legacy/protocol-sdk/src/apis/generated/premint-api-types.ts
export interface components {
  schemas: {
    /**
     * ChainName
     * @description An enumeration.
     * @enum {string}
     */
    ChainName:
      | "ETHEREUM-MAINNET"
      | "ETHEREUM-ROPSTEN"
      | "ETHEREUM-RINKEBY"
      | "ETHEREUM-GOERLI"
      | "ETHEREUM-SEPOLIA"
      | "OPTIMISM-MAINNET"
      | "OPTIMISM-GOERLI"
      | "ZORA-GOERLI"
      | "ZORA-SEPOLIA"
      | "ZORA-MAINNET"
      | "BASE-MAINNET"
      | "BASE-GOERLI"
      | "BASE-SEPOLIA"
      | "PGN-MAINNET"
      | "ARBITRUM-MAINNET"
      | "BLAST-MAINNET";
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}